Updated gradle version, added homepage and issues tags, optimized icon.
The original icon was losslessly compressed by Leanify tool and placed in the root directory of the repository. Then I made 128x128 copy and added it as a mod icon. I believe there is no reason to use (in the jar) a bigger icon since the box where it is displayed (in the Mod Menu) is small. This way the mod's size decreased greatly.